+++ LPROS/Forms_Panel_Control/PersonelController.cs
@@ -40,50 +40,57 @@ namespace LPROS.Forms_Panel_Control
         private void button_düzenle_Click(object sender, EventArgs e)
         {
             DataGridView dtg = Items.panelPersonel.dataGridview;
-            Add_Personel adPer = new Add_Personel()
+            if (dtg.Rows.Count > 0)
             {
-                isUpdate = true,
-                _SelectedTc = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["TC Kimlik Numarası"].Value.ToString(),
-                _SelectedIsim = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Adı"].Value.ToString(),
-                _SelectedSoyisim = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Soyadı"].Value.ToString(),
-                _SelectedKullaniciAdi = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Kullanıcı Adı"].Value.ToString(),
-                _SelectedSifre = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["sifre"].Value.ToString(),
-                _SelectedEposta = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["E_Posta"].Value.ToString(),
-                _SelectedAdres = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Adres"].Value.ToString(),
-                _SelectedSSKNo = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["SSK Numarası"].Value.ToString(),
-                _SelectedSSKBas = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["SSK Başlangıç Tarihi"].Value.ToString(),
-                _SelectediseGiris = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["İşe Giriş Tarihi"].Value.ToString(),
-                _SelectedCinsiyet = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Cinsiyet"].Value.ToString(),
-                _SelectedDepartman = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Departman"].Value.ToString(),
-                _SelectedYetki = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Yetki"].Value.ToString(),
-                _SelectedMaas = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Maaş"].Value.ToString(),
-                _SelectedDurumu = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Durumu"].Value.ToString(),
-                _SelectedTelefon = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Telefon Numarası"].Value.ToString(),
-                _SelectedId = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["personel_no"].Value.ToString(),
-                _SelectedGuvenlik = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Güvenlik Sorusu"].Value.ToString(),
-                _SelectedGuvenlikcvp = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Güvenlik Sorusu Cevabı"].Value.ToString(),
+                Add_Personel adPer = new Add_Personel()
+                {
+                    isUpdate = true,
+                    _SelectedTc = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["TC Kimlik Numarası"].Value.ToString(),
+                    _SelectedIsim = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Adı"].Value.ToString(),
+                    _SelectedSoyisim = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Soyadı"].Value.ToString(),
+                    _SelectedKullaniciAdi = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Kullanıcı Adı"].Value.ToString(),
+                    _SelectedSifre = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["sifre"].Value.ToString(),
+                    _SelectedEposta = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["E_Posta"].Value.ToString(),
+                    _SelectedAdres = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Adres"].Value.ToString(),
+                    _SelectedSSKNo = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["SSK Numarası"].Value.ToString(),
+                    _SelectedSSKBas = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["SSK Başlangıç Tarihi"].Value.ToString(),
+                    _SelectediseGiris = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["İşe Giriş Tarihi"].Value.ToString(),
+                    _SelectedCinsiyet = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Cinsiyet"].Value.ToString(),
+                    _SelectedDepartman = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Departman"].Value.ToString(),
+                    _SelectedYetki = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Yetki"].Value.ToString(),
+                    _SelectedMaas = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Maaş"].Value.ToString(),
+                    _SelectedDurumu = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Durumu"].Value.ToString(),
+                    _SelectedTelefon = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Telefon Numarası"].Value.ToString(),
+                    _SelectedId = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["personel_no"].Value.ToString(),
+                    _SelectedGuvenlik = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Güvenlik Sorusu"].Value.ToString(),
+                    _SelectedGuvenlikcvp = dtg.Rows[dtg.CurrentCell.RowIndex].Cells["Güvenlik Sorusu Cevabı"].Value.ToString(),
 
 
-            };
-            adPer.ShowDialog();
+                };
+                adPer.ShowDialog();
+            }
         }
 
         private void button_arama_Click(object sender, EventArgs e)
         {
             (new Search_Personel()).ShowDialog();
         }
+
         SqlConnector Sc = new SqlConnector();
         private void button_sil_Click_1(object sender, EventArgs e)
         {
-            DialogResult Dr = MessageBox.Show("Seçilen Personel Siliniyor!", "Silme İşlemi", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2);
-            if (Dr == DialogResult.OK)
+            if (Items.panelPersonel.dataGridview.Rows.Count > 0)
             {
-                DataGridView Dtg = Items.panelPersonel.dataGridview;
-
-                if (Sc.QUERY_TABLE("delete from Personel where personel_no=@parametre1", new String[] { Dtg.Rows[Dtg.CurrentCell.RowIndex].Cells["personel_no"].Value.ToString() }))
+                DialogResult Dr = MessageBox.Show("Seçilen Personel Siliniyor!", "Silme İşlemi", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2);
+                if (Dr == DialogResult.OK)
                 {
-                    MessageBox.Show("Silme İşlemi Başarılı!", "Silme İşlemi", MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
-                    Items.panelPersonel.dataGridview.DataSource = Sc.GET_DATATABLE(SqlConnector.TablePersonel);
+                    DataGridView Dtg = Items.panelPersonel.dataGridview;
+
+                    if (Sc.QUERY_TABLE("delete from Personel where personel_no=@parametre1", new String[] { Dtg.Rows[Dtg.CurrentCell.RowIndex].Cells["personel_no"].Value.ToString() }))
+                    {
+                        MessageBox.Show("Silme İşlemi Başarılı!", "Silme İşlemi", MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
+                        Items.panelPersonel.dataGridview.DataSource = Sc.GET_DATATABLE(SqlConnector.TablePersonel);
+                    }
                 }
             }
         }
